requirements.txt
gspread
oauth2client
google-generativeai